Crate rendy_mesh

This crates provides means to deal with vertex buffers and meshes.

Attribute and VertexFormat allow vertex structure to declare semantics. Mesh can be created from typed vertex structures and provides mechanism to bind vertex attributes required by shader interface.

Structs

AttrGenIter
An iterator adapter that generates a list of attributes with given formats and names

AttrUuid
A unique identifier for vertex attribute of given name, format and array index.

Attribute
Vertex attribute type.

Color
Type for color attribute of vertex

Incompatible
Error type returned by Mesh::bind in case of mesh's vertex buffers are incompatible with requested vertex formats.

IndexBuffer
Index buffer with it's type

Mesh
Single mesh is a collection of buffer ranges that provides available attributes. Usually exactly one mesh is used per draw call.

MeshBuilder
Generics-free mesh builder.

Model
Full vertex transformation attribute. Typically provided on per-instance basis. It takes 4 attribute locations.

Normal
Type for texture coord attribute of vertex

PosColor
Vertex format with position and RGBA8 color attributes.

PosColorNorm
Vertex format with position, color and normal attributes.

PosNorm
Vertex format with position and normal attributes.

PosNormTangTex
Vertex format with position, normal, tangent, and UV texture coordinate attributes.

PosNormTex
Vertex format with position, normal and UV texture coordinate attributes.

PosTex
Vertex format with position and UV texture coordinate attributes.

Position
Type for position attribute of vertex.

Tangent
Type for tangent attribute of vertex. W represents handedness and should always be 1 or -1

TexCoord
Type for texture coord attribute of vertex

VertexBufferLayout
Vertex buffer with it's format

VertexFormat
Vertex format contains information to initialize graphics pipeline Attributes must be sorted by offset.

Enums

Indices
Abstracts over two types of indices and their absence.

Traits

AsAttribute
Trait for vertex attributes to implement

AsAttributes
Represent types that can be interpreted as list of vertex attributes.

AsVertex
Trait implemented by all valid vertex formats.

Functions

attribute_uuid
Retreive a unique identifier for vertex attribute of given name, format and array index.