Struct rendy_mesh::MeshBuilder

pub struct MeshBuilder<'a> { /* fields omitted */ }

Generics-free mesh builder.

Implementations

impl<'a> MeshBuilder<'a>

pub fn new() -> Self

Create empty builder.

pub fn into_owned(self) -> MeshBuilder<'static>

Convert builder into fully owned type. This forces internal vertex and index buffers to be cloned, which allows borrowed source buffers to be released.

pub fn with_indices<I>(self, indices: I) -> Self where
I: Into<Indices<'a>>,

Set indices buffer to the MeshBuilder

pub fn set_indices<I>(&mut self, indices: I) -> &mut Self where
I: Into<Indices<'a>>,

Set indices buffer to the MeshBuilder

pub fn with_vertices<V, D>(self, vertices: D) -> Self where
V: AsVertex + 'a,
D: Into<Cow<'a, [V]>>,

Add another vertices to the MeshBuilder

pub fn add_vertices<V, D>(&mut self, vertices: D) -> &mut Self where
V: AsVertex + 'a,
D: Into<Cow<'a, [V]>>,

Add another vertices to the MeshBuilder

pub fn with_prim_type(self, prim: Primitive) -> Self

Sets the primitive type of the mesh.

By default, meshes are constructed as triangle lists.

pub fn set_prim_type(&mut self, prim: Primitive) -> &mut Self

Sets the primitive type of the mesh.

By default, meshes are constructed as triangle lists.

pub fn build<B>(
&self,
queue: QueueId,
factory: &Factory<B>
) -> Result<Mesh<B>, UploadError> where
B: Backend,

Builds and returns the new mesh.

A mesh expects all vertex buffers to have the same number of elements. If those are not equal, the length of smallest vertex buffer is selected, effectively discaring extra data from larger buffers.

Note that contents of index buffer is not validated.